Fix require cycle by moving setPushNotificationOptInStatus out of User
